refactor(tools): extract todo persistence helpers into microclaw-tools
