+++ crates/microclaw-tools/Cargo.toml
@@ -10,6 +10,7 @@ async-trait = "0.1"
 microclaw-core = { path = "../microclaw-core" }
 reqwest = { version = "0.12", features = ["json", "blocking"] }
 serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 tokio = { version = "1", features = ["full"] }
 tracing = "0.1"
 urlencoding = "2"
